backend/requirements.txt
# Vanguard pipeline
pyyaml>=6.0
requests>=2.28.0
httpx>=0.24.0
openai>=1.0.0
python-dotenv>=1.0.0
# Name-to-search pipeline (production-grade NLP)
regex>=2023.0
rapidfuzz>=3.0
wordfreq>=3.0
unidecode>=1.3.0
spacy>=3.5.0
pytest>=7.0
